refactor: context를 멤버변수로 사용하지 않고 클릭리스너 사용할 수 있도록 수정
