+++ app/src/main/java/woowacourse/movie/movielist/AdViewHolder.kt
@@ -6,14 +6,14 @@ import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
 import woowacourse.movie.dto.AdDto
 
-class AdViewHolder(private val view: View, private val listener: OnClickListener<AdDto>) :
+class AdViewHolder(private val view: View, private val onAdClickListener: OnClickListener<AdDto>) :
     RecyclerView.ViewHolder(view) {
     val adImg = itemView.findViewById<ImageView>(R.id.ad)
 
     fun bind(item: AdDto) {
         adImg.setImageResource(item.adImage)
         adImg.setOnClickListener {
-            listener.onClick(item)
+            onAdClickListener.onClick(item)
         }
     }
 }
