+++ app/src/main/java/woowacourse/movie/movielist/MainActivity.kt
@@ -1,16 +1,18 @@
 package woowacourse.movie.movielist
 
+import android.content.Intent
+import android.net.Uri
 import android.os.Bundle
 import androidx.appcompat.app.AppCompatActivity
 import androidx.recyclerview.widget.RecyclerView
 import woowacourse.movie.R
 import woowacourse.movie.dto.AdDto
+import woowacourse.movie.dto.MovieDto
 import woowacourse.movie.dto.MovieDummy
+import woowacourse.movie.moviedetail.MovieDetailActivity
 
 class MainActivity : AppCompatActivity() {
 
-    private val onMovieClickListener by lazy { OnMovieClickListener(this) }
-    private val onAdClickListener by lazy { OnAdClickListener(this) }
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
@@ -23,11 +25,36 @@ class MainActivity : AppCompatActivity() {
         val movieRVAdapter = MovieRVAdapter(
             MovieDummy.movieDatas,
             AdDto.getAdData(),
-            onMovieClickListener,
-            onAdClickListener,
         )
 
         movieRV.adapter = movieRVAdapter
+
+        onMovieItemClickListener(movieRVAdapter)
+        onAdItemClickListener(movieRVAdapter)
+
         movieRVAdapter.notifyDataSetChanged()
+    }
+
+    private fun onMovieItemClickListener(adapter: MovieRVAdapter) {
+        adapter.itemMovieClick = object : OnClickListener<MovieDto> {
+            override fun onClick(item: MovieDto) {
+                val intent = Intent(this@MainActivity, MovieDetailActivity::class.java)
+                intent.putExtra(MOVIE_KEY, item)
+                startActivity(intent)
+            }
+        }
+    }
+
+    private fun onAdItemClickListener(adapter: MovieRVAdapter) {
+        adapter.itemAdClick = object : OnClickListener<AdDto> {
+            override fun onClick(item: AdDto) {
+                val intent = Intent(Intent.ACTION_VIEW, Uri.parse(item.url))
+                startActivity(intent)
+            }
+        }
+    }
+
+    companion object {
+        private const val MOVIE_KEY = "movie"
     }
 }
